browser/components/resistfingerprinting/test/mochitest/test_reduce_time_precision.html

<!DOCTYPE HTML>
<html>
<!--
Tor bug
https://trac.torproject.org/projects/tor/ticket/1517
-->
<head>
  <meta charset="utf-8">
  <title>Test for Tor Bug 1517 and Mozilla Bug 1424341</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SpawnTask.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://trac.torproject.org/projects/tor/ticket/1517">Tor Bug 1517</a>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1424341">Mozilla Bug 1424341</a>

<!-- Canvas for testing 'currentTime' -->
<canvas id="test-canvas" width="100" height="100"></canvas>

<!-- The main testing script -->
<script type="application/javascript">
  SimpleTest.requestFlakyTimeout("testing JS time-based fingerprinting");

  // Prepare for test of AudioContext.currentTime
  let audioContext = new AudioContext();
  // Prepare for test of CanvasStream.currentTime
  let canvas = document.getElementById("test-canvas");
  let context = canvas.getContext("2d");
  context.fillText("test", 20, 20);
  let canvasStream = canvas.captureStream(25);

  // Known ways to generate time stamps, in milliseconds
  const timeStampCodes = [
    "performance.now()",
    "new Date().getTime()",
    "new Event(\"\").timeStamp",
    "new File([], \"\").lastModified",
    "new File([], \"\").lastModifiedDate.getTime()",
  ];
  // These are measured in seconds, so we need to scale them up
  var timeStampCodesDOM = timeStampCodes.concat([
    "audioContext.currentTime * 1000",
    "canvasStream.currentTime * 1000",
  ]);

  let isRounded = (x, expectedPrecision) => {
    let rounded = (Math.floor(x / expectedPrecision) * expectedPrecision);
    // First we do the perfectly normal check that should work just fine
    if (rounded === x || x === 0)
      return true;

    // When we're diving by non-whole numbers, we may not get perfect
    // multiplication/division because of floating points
    if (Math.abs(rounded - x + expectedPrecision) < .0000001) {
      return true;
    } else if (Math.abs(rounded - x) < .0000001) {
      return true;
    }

    // Then we handle the case where you're sub-millisecond and the timer is not
    // We check that the timer is not sub-millisecond by assuming it is not if it
    // returns an even number of milliseconds
    if (expectedPrecision < 1 && Math.round(x) == x) {
      if (Math.round(rounded) == x) {
        return true;
      }
    }

    ok(false, "Looming Test Failure, Additional Debugging Info: Expected Precision: " + expectedPrecision + " Measured Value: " + x +
      " Rounded Vaue: " + rounded + " Fuzzy1: " + Math.abs(rounded - x + expectedPrecision) +
      " Fuzzy 2: " + Math.abs(rounded - x));

    return false;
  };

  // ================================================================================================
  // ================================================================================================
  async function checkWorker(worker, prefname, expectedPrecision) {
    // The child worker will send the results back.
    let checkWorkerTimeStamps = () => new Promise(function(resolve) {
      let onMessage = function(event) {
        worker.removeEventListener("message", onMessage);

        let timeStamps = event.data;
        for (let i = 0; i < timeStampCodes.length; i++) {
          let timeStamp = timeStamps[i];
          ok(isRounded(timeStamp, expectedPrecision),
            "pref: " + prefname + " - '" +
             "'" + timeStampCodes[i] +
             "' should be rounded to nearest " + expectedPrecision + " ms in workers; saw " +
             timeStamp);
        }
        resolve();
      };
      worker.addEventListener("message", onMessage);
    });

    // Send the codes to its child worker.
    worker.postMessage(timeStampCodes);

    // First, check the child's results.
    await checkWorkerTimeStamps();
    // Then, check the grandchild's results.
    await checkWorkerTimeStamps();

    worker.terminate();
  }

  async function testWorker(resistFingerprinting, reduceTimerPrecision, expectedPrecision) {
    let prefname = "";
    prefname += resistFingerprinting ? "privacy.resistFingerprinting " : "";
    prefname += reduceTimerPrecision ? "privacy.reduceTimerPrecision " : "";
    // Create one worker before setting the pref, and one after, in order to
    // check that the resolution is updated whether or not the worker was
    // already started
    let worker1 = new Worker("worker_child.js");
    await SpecialPowers.pushPrefEnv({
      "set": [["privacy.resistFingerprinting", resistFingerprinting],
              ["privacy.reduceTimerPrecision", reduceTimerPrecision],
              ["privacy.resistFingerprinting.reduceTimerPrecision.microseconds", expectedPrecision * 1000]
              ]});

    // No matter what we set the precision to, if we're in ResistFingerprinting mode
    // we use the larger of the precision pref and the constant 100ms
    if (resistFingerprinting) {
      expectedPrecision = expectedPrecision < 100 ? 100 : expectedPrecision;
    }

    let worker2 = new Worker("worker_child.js");
    // Allow ~550 ms to elapse, so we can get non-zero
    // time values for all elements.
    await new Promise(resolve => window.setTimeout(resolve, 550));
    await checkWorker(worker1, prefname, expectedPrecision);
    await checkWorker(worker2, prefname, expectedPrecision);
  }

  add_task(async function testWorkerRFP() {
    await testWorker(true, false, 100);
    await testWorker(true, false, 13);
    await testWorker(true, false, .13);
  });

  add_task(async function testWorkerRTP() {
    await testWorker(false, true, 100);
    await testWorker(false, true, 13);
    await testWorker(false, true, .13);
  });

  // ================================================================================================
  // ================================================================================================
  async function testDOM(resistFingerprinting, reduceTimerPrecision, expectedPrecision) {
    let prefname = "";
    prefname += resistFingerprinting ? "privacy.resistFingerprinting " : "";
    prefname += reduceTimerPrecision ? "privacy.reduceTimerPrecision " : "";

    await SpecialPowers.pushPrefEnv({
      "set": [["privacy.resistFingerprinting", resistFingerprinting],
              ["privacy.reduceTimerPrecision", reduceTimerPrecision],
              ["privacy.resistFingerprinting.reduceTimerPrecision.microseconds", expectedPrecision * 1000]
              ]});

    // No matter what we set the precision to, if we're in ResistFingerprinting mode
    // we use the larger of the precision pref and the constant 100ms
    if (resistFingerprinting) {
      expectedPrecision = expectedPrecision < 100 ? 100 : expectedPrecision;
    }

    // Loop through each timeStampCode, evaluate it,
    // and check if it is rounded
    for (let timeStampCode of timeStampCodesDOM) {
      let timeStamp = eval(timeStampCode);
      ok(isRounded(timeStamp, expectedPrecision),
        "pref: " + prefname + " - '" +
         "'" + timeStampCode +
         "' should be rounded to nearest " +
         expectedPrecision + " ms; saw " +
         timeStamp);
    }
  }

  add_task(async function testDOMRFP() {
    await testDOM(true, true, 100);
    await testDOM(true, false, 13);
    await testDOM(true, false, .13);
  });

  add_task(async function testDOMRTP() {
    await testDOM(false, true, 100);
    await testDOM(false, true, 13);
    await testDOM(false, true, .13);
  });


</script>


</body>
</html>
